pyproject.toml
[project]
name = "agentic-patterns"
version = "1.0.0"
description = "Comprehensive implementation guide for 35 agentic design patterns with Google ADK and CrewAI"
authors = [
    {name = "Matteo Gazzurelli", email = "matteo@gazzurelli.com"}
]
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.12"
keywords = ["ai", "agents", "agentic-patterns", "google-adk", "crewai", "llm", "multi-agent"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core AI frameworks
    "google-cloud-aiplatform>=1.128.0",
    "google-adk>=1.18.0",
    "crewai>=1.5.0",
    "crewai-tools>=1.5.0",
    # LLM providers
    "openai>=2.8.1",
    "anthropic>=0.74.0",
    # Vector stores and embeddings
    "chromadb>=1.1.1",
    "sentence-transformers>=5.1.2",
    "faiss-cpu>=1.13.0",
    # Data processing
    "pandas>=2.3.3",
    "numpy>=2.3.5",
    "pydantic>=2.12.4",
    # Utilities
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "rich>=14.2.0",
    "loguru>=0.7.3",
    # Web and APIs
    "httpx>=0.28.1",
    "requests>=2.32.5",
    # MCP and protocols
    "mcp>=1.21.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "black>=25.11.0",
    "ruff>=0.14.5",
    "mypy>=1.18.2",
    "ipython>=9.7.0",
    "jupyter>=1.1.1",
]

benchmark = [
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    "plotly>=6.5.0",
    "psutil>=7.1.3",
]

docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=0.30.1",
]

all = [
    "agentic-patterns[dev,benchmark,docs]",
]

[project.urls]
Homepage = "https://github.com/gazzumatteo/agentic-patterns"
Documentation = "https://github.com/gazzumatteo/agentic-patterns/docs"
Repository = "https://github.com/gazzumatteo/agentic-patterns"
"Bug Tracker" = "https://github.com/gazzumatteo/agentic-patterns/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared-utilities"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
    "pytest-mock>=3.15.1",
    "mypy>=1.18.2",
    "ipython>=9.7.0",
    "jupyter>=1.1.1",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = [
    "--verbose",
    "--cov=.",
    "--cov-report=html",
    "--cov-report=term-missing",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
